additional refactoring: slower
